Update README.md
Included updated client type when storing your client in your secret vault. Previously it was only designed for just API, now you will have the option to store a service account (Legacy) and the vault will be able to identify which one to authenticate with and generate a auth token.